LayoutTests/http/tests/security/frameNavigation/navigation-ALLOWED-sibling-same-parent.html

<html>
<head>
<script src="resources/lib.js"></script>
<script>
navUtils.set_frame_count(10);

navUtils.onready = function () {
  navUtils.log('Navigating frames.');

  navUtils.set_frame_count(5);

  navUtils.doNavigate(frames['active0'], 'location', 'window.top.frames[0]');
  navUtils.doNavigate(frames['active1'], 'open', 'child1');
  navUtils.doNavigate(frames['active2'], 'form', 'child2');
  navUtils.doNavigate(frames['active3'], 'hyperlink', 'child3');
  navUtils.doNavigate(frames['active4'], 'plugin', 'child4');
};

navUtils.ondone = function () {
  navUtils.logWindowCount();
  navUtils.log('Done.');
  if (window.layoutTestController)
    layoutTestController.notifyDone();
};
</script>
</head>
<body>
<div id='console'></div>
<iframe name="child0" src="http://localhost:8000/security/frameNavigation/resources/ready.html"></iframe>
<iframe name="child1" src="http://localhost:8000/security/frameNavigation/resources/ready.html"></iframe>
<iframe name="child2" src="http://localhost:8000/security/frameNavigation/resources/ready.html"></iframe>
<iframe name="child3" src="http://localhost:8000/security/frameNavigation/resources/ready.html"></iframe>
<iframe name="child4" src="http://localhost:8000/security/frameNavigation/resources/ready.html"></iframe>

<iframe name="active0" src="http://127.0.0.1:8000/security/frameNavigation/resources/ready.html"></iframe>
<iframe name="active1" src="http://127.0.0.1:8000/security/frameNavigation/resources/ready.html"></iframe>
<iframe name="active2" src="http://127.0.0.1:8000/security/frameNavigation/resources/ready.html"></iframe>
<iframe name="active3" src="http://127.0.0.1:8000/security/frameNavigation/resources/ready.html"></iframe>
<iframe name="active4" src="http://127.0.0.1:8000/security/frameNavigation/resources/ready.html"></iframe>
</body>
</html>
