LayoutTests/http/tests/security/frameNavigation/navigation-ALLOWED-sibling-same-parent-expected.txt

Navigating frames.
Window count: 1
Done.
          

--------
Frame: 'child0'
--------
127.0.0.1

Navigation done.

--------
Frame: 'child1'
--------
127.0.0.1

Navigation done.

--------
Frame: 'child2'
--------
127.0.0.1

Navigation done.

--------
Frame: 'child3'
--------
127.0.0.1

Navigation done.

--------
Frame: 'child4'
--------
127.0.0.1

Navigation done.

--------
Frame: 'active0'
--------
127.0.0.1:8000

Waiting for navigation commands.

--------
Frame: 'active1'
--------
127.0.0.1:8000

Waiting for navigation commands.

--------
Frame: 'active2'
--------
127.0.0.1:8000

Waiting for navigation commands.


--------
Frame: 'active3'
--------
127.0.0.1:8000

Waiting for navigation commands.

--------
Frame: 'active4'
--------
127.0.0.1:8000

Waiting for navigation commands. 
